+++ test/measure/area_test.dart
@@ -1,5 +1,5 @@
-import 'package:geodart/features.dart';
-import 'package:geodart/measure.dart';
+import 'package:geodart/geometries.dart';
+import 'package:geodart/measurements.dart';
 import 'package:test/test.dart';
 
 void main() {