+++ CHANGELOG.md
@@ -1,8 +1,14 @@
-## 0.0.1
+## 0.1.0
 
-- Initial version.
+- Remove installation information from the README.md
+- Add conversion library, with 1d conversions, with tests
 
 ## 0.0.2
 
 - Changed pubspec information
 - Un-ignored docs and examples
+
+## 0.0.1
+
+- Initial version.
+